move bitMask64 to test file
